Refactor | Client | Exract components to separate directories
