+++ client/src/views/ExpenseEdit/ExpenseEdit.component.jsx
@@ -4,8 +4,8 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import Select from 'react-select';
 
-import Button from '../../components/Button';
-import ContentWrapper from '../../components/ContentWrapper';
+import Button from '../../components/Button/Button.component';
+import ContentWrapper from '../../components/ContentWrapper/ContentWrapper.component';
 import {
   customStyleSelect,
   Form,
