+++ client/src/views/TripAdd/TripAdd.component.jsx
@@ -7,8 +7,8 @@ import DatePicker from 'react-datepicker';
 import { connect } from 'react-redux';
 import Select from 'react-select';
 
-import Button from '../../components/Button';
-import ContentWrapper from '../../components/ContentWrapper';
+import Button from '../../components/Button/Button.component';
+import ContentWrapper from '../../components/ContentWrapper/ContentWrapper.component';
 import {
   customStyleSelect,
   DateInput,
