+++ client/src/views/UserProfile/UserProfile.component.jsx
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
 
-import ContentWrapper from '../../components/ContentWrapper';
+import ContentWrapper from '../../components/ContentWrapper/ContentWrapper.component';
 import { H3, InnerContainer, LinkButtonSmall } from '../../components/styled';
 import getToken from '../../utils/getToken';
 import { theme } from '../../utils/theme';
